# openapi-v1.json
{
  "components": {
    "schemas": {
      "AccumulatedDoseResult": {
        "additionalProperties": true,
        "description": "Returned by ``deform_service.run_dose_accumulation`` (``report.as_dict()``).\n\nThe exact keys depend on the GenDoseCalc version, so ``extra=\"allow\"``\npasses unknown fields through unchanged.",
        "properties": {
          "accumulated_max_gy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accumulated Max Gy"
          },
          "accumulated_mean_gy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accumulated Mean Gy"
          },
          "n_states": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N States"
          },
          "total_elapsed_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Elapsed S"
          },
          "total_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Weight"
          }
        },
        "title": "AccumulatedDoseResult",
        "type": "object"
      },
...
      "ContourReview": {
        "properties": {
          "fraction_index": {
            "title": "Fraction Index",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "structure_id": {
            "title": "Structure Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "fraction_index",
          "structure_id",
          "status"
        ],
        "title": "ContourReview",
        "type": "object"
      },
      "ContourStatusBody": {
        "properties": {
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ContourStatusBody",
        "type": "object"
      },
...
      "Fraction": {
        "properties": {
          "dataset_item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dataset Item Id"
          },
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "machine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Machine"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          }
        },
        "required": [
          "index",
          "label",
          "session_id"
        ],
        "title": "Fraction",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RegistrationMetric": {
        "properties": {
          "approved": {
            "default": false,
            "title": "Approved",
            "type": "boolean"
          },
          "fraction_index": {
            "title": "Fraction Index",
            "type": "integer"
          },
          "mean_dice": {
            "title": "Mean Dice",
            "type": "number"
          },
          "rms_surface_distance_mm": {
            "title": "Rms Surface Distance Mm",
            "type": "number"
          }
        },
        "required": [
          "fraction_index",
          "rms_surface_distance_mm",
          "mean_dice"
        ],
        "title": "RegistrationMetric",
        "type": "object"
      },
      "RegistrationSummary": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "metrics": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/RegistrationMetric"
            },
            "title": "Metrics",
            "type": "array"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "session_id",
          "job_id",
          "status"
        ],
        "title": "RegistrationSummary",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Axioma Sidecar",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/adaptive/contours/{contour_id}/status": {
      "patch": {
        "operationId": "update_contour_status_api_v1_adaptive_contours__contour_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "contour_id",
            "required": true,
            "schema": {
              "title": "Contour Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContourStatusBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Contour Status"
      }
    },
    "/api/v1/adaptive/sessions/{session_id}/dose-accumulation": {
      "get": {
        "operationId": "get_dose_accumulation_api_v1_adaptive_sessions__session_id__dose_accumulation_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccumulatedDoseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Dose Accumulation"
      }
    },
    "/api/v1/adaptive/sessions/{session_id}/registrations": {
      "get": {
        "description": "Return stored registration results for this session.\n\nResults are written by the 'register' job worker after the engine completes.\nThis endpoint only reads \u2014 it never invokes the engine.",
        "operationId": "get_registrations_api_v1_adaptive_sessions__session_id__registrations_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RegistrationMetric"
                  },
                  "title": "Response Get Registrations Api V1 Adaptive Sessions  Session Id  Registrations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Registrations"
      }
    },
...
    "/api/v1/registrations/{registration_id}": {
      "get": {
        "description": "Return one completed registration job without invoking the engine.",
        "operationId": "get_registration_api_v1_registrations__registration_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "title": "Registration Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Registration"
      }
    },
    "/api/v1/registrations/{registration_id}/contours/{roi}/accept": {
      "post": {
        "description": "Persist acceptance for a registration ROI; repeated acceptance is idempotent.",
        "operationId": "accept_contour_api_v1_registrations__registration_id__contours__roi__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "registration_id",
            "required": true,
            "schema": {
              "title": "Registration Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "roi",
            "required": true,
            "schema": {
              "title": "Roi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContourReview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accept Contour"
      }
    },
...
    "/api/v1/sessions/{session_id}/fractions": {
      "get": {
        "description": "Return ordered fraction groups already discovered in this session.",
        "operationId": "get_fractions_api_v1_sessions__session_id__fractions_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Fraction"
                  },
                  "title": "Response Get Fractions Api V1 Sessions  Session Id  Fractions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Fractions"
      }
    },
...
  }
}